refactor(error_tracking): move remaining api endpoints to presentation
