+++ products/error_tracking/backend/presentation/__init__.py
@@ -1,15 +1,57 @@
 from .serializers import (
+    ErrorTrackingAssignmentRuleSerializer,
     ErrorTrackingExternalReferenceIntegrationSerializer,
     ErrorTrackingExternalReferenceSerializer,
     ErrorTrackingFingerprintSerializer,
+    ErrorTrackingGroupingRuleSerializer,
+    ErrorTrackingIssueFullSerializer,
+    ErrorTrackingIssuePreviewSerializer,
+    ErrorTrackingReleaseSerializer,
+    ErrorTrackingSpikeDetectionConfigSerializer,
+    ErrorTrackingSpikeEventSerializer,
+    ErrorTrackingStackFrameSerializer,
+    ErrorTrackingSuppressionRuleSerializer,
+    ErrorTrackingSymbolSetSerializer,
 )
-from .views import ErrorTrackingExternalReferenceViewSet, ErrorTrackingFingerprintViewSet, GitProviderFileLinksViewSet
+from .views import (
+    ErrorTrackingAssignmentRuleViewSet,
+    ErrorTrackingExternalReferenceViewSet,
+    ErrorTrackingFingerprintViewSet,
+    ErrorTrackingGroupingRuleViewSet,
+    ErrorTrackingIssueViewSet,
+    ErrorTrackingReleaseViewSet,
+    ErrorTrackingSpikeDetectionConfigViewSet,
+    ErrorTrackingSpikeEventViewSet,
+    ErrorTrackingStackFrameViewSet,
+    ErrorTrackingSuppressionRuleViewSet,
+    ErrorTrackingSymbolSetViewSet,
+    GitProviderFileLinksViewSet,
+)
 
 __all__ = [
-    "ErrorTrackingFingerprintSerializer",
-    "ErrorTrackingFingerprintViewSet",
+    "ErrorTrackingAssignmentRuleSerializer",
+    "ErrorTrackingAssignmentRuleViewSet",
     "ErrorTrackingExternalReferenceIntegrationSerializer",
     "ErrorTrackingExternalReferenceSerializer",
     "ErrorTrackingExternalReferenceViewSet",
+    "ErrorTrackingFingerprintSerializer",
+    "ErrorTrackingFingerprintViewSet",
+    "ErrorTrackingGroupingRuleSerializer",
+    "ErrorTrackingGroupingRuleViewSet",
+    "ErrorTrackingIssueFullSerializer",
+    "ErrorTrackingIssuePreviewSerializer",
+    "ErrorTrackingIssueViewSet",
+    "ErrorTrackingReleaseSerializer",
+    "ErrorTrackingReleaseViewSet",
+    "ErrorTrackingSpikeDetectionConfigSerializer",
+    "ErrorTrackingSpikeDetectionConfigViewSet",
+    "ErrorTrackingSpikeEventSerializer",
+    "ErrorTrackingSpikeEventViewSet",
+    "ErrorTrackingStackFrameSerializer",
+    "ErrorTrackingStackFrameViewSet",
+    "ErrorTrackingSuppressionRuleSerializer",
+    "ErrorTrackingSuppressionRuleViewSet",
+    "ErrorTrackingSymbolSetSerializer",
+    "ErrorTrackingSymbolSetViewSet",
     "GitProviderFileLinksViewSet",
 ]
